rename to spt_compute; move process lock to separate file
