+++ spt_compute/__init__.py
@@ -1,0 +1,2 @@
+from .ecmwf_rapid_process import run_ecmwf_rapid_process
+from .process_lock import reset_lock_info_file
