+++ spt_compute/autorapid_process.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ##
 ##  autorapid_process.py
-##  spt_ecmwf_autorapid_process
+##  spt_compute
 ##
 ##  Created by Alan D. Snow.
 ##  Copyright © 2015-2016 Alan D Snow. All rights reserved.
