+++ spt_compute/imports/ftp_ecmwf_download.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ##
 ##  ftp_ecmwf_download.py
-##  spt_ecmwf_autorapid_process
+##  spt_compute
 ##
 ##  Created by Alan D. Snow.
 ##  Copyright © 2015-2016 Alan D Snow. All rights reserved.
