+++ spt_compute/imports/helper_functions.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ##
 ##  helper_functions.py
-##  spt_ecmwf_autorapid_process
+##  spt_compute
 ##
 ##  Created by Alan D. Snow.
 ##  Copyright © 2015-2016 Alan D Snow. All rights reserved.
@@ -13,6 +13,7 @@ import os
 import re
 from shutil import rmtree
 import sys
+
 
 #----------------------------------------------------------------------------------------
 # HELPER FUNCTIONS
