--- spt_process/__init__.py
@@ -1,1 +1,0 @@
-from ecmwf_rapid_process import reset_lock_info_file, run_ecmwf_rapid_process
